Added new validation classes to make passing messages through to the process functions much easier. Validation now comprises of a parent validator class which makes use of a validation class for an individual field validation. The parent class has useful methods like allFieldsAreEmpty() to ensure that a stage that hasn't had anything filled in will not produce errors.
Also removed tooltips from the field titles and instead added a new
element to the field arrays called #help_link. If a help link is found
during label rendering the help tip will be shown. This allows us to
show the question title on validation error in a validation header area.