Fix for PDF download timeout.
The problem occured due to base64 encoding the image url for the flow
diagram and then using that encoded url in the theme tpl. Instead I've
changed the image handling to use an image derivative to get a maximum
width of 2048 pixels and used that derivative image in the PDF output.