12348 Fixed image rendering issue in PDF
Due to a bug in file entity there used to be no checking of file
permissions for private files. Once the bug was fixed private files were
no longer rendering in the PDF or as previews. Even with the correct
permissions to the PDF was unable to render the image due to the session
calling the image being anonymous. To get around this problem I used a
method of copying the private file to the temp directory, renaming it
with an MD5 for safety and then using that derivative image in the PDF.
Once the PDF rendering is complete the temporary file is deleted.