12348 Fixed bug in PDF output with wrong IID
Due to the IID not being set soon enough in the PDF output code in some
situations the IID from another study (owned by the same person), if
previously viewed, could be the IID used by the PDF output in some of
its functions. This would result in mismatches between stored data and
viewed data which would manifest as missing text in the PDF. Setting the
IID slightly earlier fixes the problem.