Added a business name field and rendered on PDF
Made an update hook which creates a business name field and an
associated instance.
Created a business class with getter and setter methods for getting the
business name. This is used in the pdf include to render out the
name if set. Also includes some error checking if the user does not have
a profile.